TANGERINE-1042: continued work on batch; use dummy URL postBatchGrid.htm to access new Ext Batch View
git-svn-id: svn://svn.orangeleap.com/tangerine/tangerine/trunk@8794 0f581344-5646-3a48-8ded-616bfea576c6

+++ src/com/orangeleap/tangerine/domain/Segmentation.java
@@ -10,13 +10,13 @@ public class Segmentation implements Serializable {
     private Long id;
     private String name;
     private String description;
-    private int count;
+    private long count;
     private Date lastRunDate;
     private String lastRunByUser;
 
     public Segmentation() { }
 
-    public Segmentation(Long id, String name, String description, int count, Date lastRunDate, String lastRunByUser) {
+    public Segmentation(Long id, String name, String description, long count, Date lastRunDate, String lastRunByUser) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -49,11 +49,11 @@ public class Segmentation implements Serializable {
         this.description = description;
     }
 
-    public int getCount() {
+    public long getCount() {
         return count;
     }
 
-    public void setCount(int count) {
+    public void setCount(long count) {
         this.count = count;
     }
